[MAIN] Added container and network as farms dependencies in setup
Now farms_(container, network and pylog) will be directly installed as
dependencies. With this we can move deprecated Neuromechfly copies of
network and container.

WARNING : We need to make sure that some files changed in network are
reflected in the main repository appropriately. For example,
agnostic_controller.py was changed locally!